move most of the code so far to library lib/mercurial.b (while creating a library interface).
add lots of code, many adts.
Repo is a repository (mostly a path and some flags).
Revlog is an open revlog file (e.g. 00changelog.i, or a data/mkfile.i).  there is no support yet for .d files.
Change is a change as occurs in a changelog
Manifest is a manifest listing (file names & nodeids, Manifestfiles)
Nodeid is a sha1 hash
Entry is a index entry from .i files

getting other revisions than the first is currently broken.  will be fixed next.