be a bit more efficient in reconstructing versions from a base+delta's.
the old code would take the base, apply one patch, resulting in the
base for the next patch.  this involved quite some copying for each
step (far too much actually).  the intention was that all patches
would be merged into a big patch, then applied to the single base.
but merging patches is relatively tricky.  instead, this adds code
to apply a patch to a base that is not a single contiguous buffer,
but a list of buffers.  the result is again a list of buffers.  so
the process is:

1. wrap the base buffer into the list (with just one element)
2. for each patch, apply to the list of buffers, resulting in a new
   list of buffers
3. flatten the list of buffers, returning the final buffer.

the intermediate buffers are sliced arrays from the initial base,
and from the buffers in the patches.  that doesn't involve any
copying, just array object creation.  only at the end, while
flattening, is all the memory copying done.  there is no way around
that.

a next version will make applying the changes a bit more efficient.
at least this is a simple version that works.