make applying multiple patches a bit more efficient.
when doing

	(source group + patch) -> destination group

and taking a buffer (start, end) from source group, we no longer
seek to "start" from the beginning.  instead we keep "source group"
at the offset of the previous operation, removing the need to walk
through the list of buffers all the time.

we also keep track of the length of the group while constructing
it.  not by calculating it at the end.