greatly simplify and speed up mercurial library.
we no longer try to be smart and fail miserably.  instead we just
read the entire index file when opening a revlog, and keep it in
memory.  additionally we keep a cache of decompressed buffers that
we recently read.

also a bugfix to hg/websrv, where "branches" didn't return the right
base if it was the node requested.