fix clone for inferno-os.  a few other nits.
the problem was that for changegroup calculation the nullnode was
handled as if it was the first revision.  that assumption is bogus.
also set the manifest link properly (ie not the one in the Entry,
but for the actual revision we it belongs to)

also use hgwire debug flag in hg/fs (for its wire file).

also disable building appl/cmd/^(test cgi).  accidentially enabled
in previous commit.