welcome hg/update and hg/identify.  update hg/status.  and other changes.
note: hg/update & hg/identify are not yet ready for use, there will be bugs.

hg/update will change the files in the checked out directory to the
one specified on the command-line, or "tip" if none specified.

hg/identify prints information about the current checked out copy,
including nodeid, tags and the branch if non-default.

this also starts making use of util0.m.  and has a usable mkconfig
delivered with it.