give mercurial lib's Repo and Revlog functions that do not return tuples with an error string component but instead raise an error.
this reduces the amount of error string checking, and allows better
use of type inference of new variables.  the new functions start
with an x, to show the difference (though type errors will be caught
on wrong usage).  we keep the old functions for now, some are used.
on average, this gives smaller code, but with less clear error
messages.